Move code for configuring base voltage color to a dedicated module
