+++ base-voltage-color/src/main/java/com/powsybl/basevoltage/BaseVoltagesConfig.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.substationdiagram.svg.config;
+package com.powsybl.basevoltage;
 
 import java.util.List;
