+++ base-voltage-color/src/test/resources/base-voltages.yml
